Refactor with test updates
