+++ viewer/src/components/Dashboard.test.js
@@ -1,5 +1,6 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import Dashboard from './Dashboard';
+import { renderWithProviders } from '../utils/test-utils';
 
 const getProject = (items) => {
   return {
@@ -13,14 +14,10 @@ const getProject = (items) => {
   }
 };
 
-const fetchTraces = () => {
-  return []
-}
-
 test('renders dashboard chart', async () => {
   const project = getProject([{ width: 1, chart: { name: "chart_name", traces: [] } }])
 
-  render(<Dashboard project={project} fetchTraces={fetchTraces} dashboardName={'dashboard'} />)
+  renderWithProviders(<Dashboard project={project} dashboardName={'dashboard'} />)
 
   await waitFor(() => {
     expect(screen.getByTestId('dashboard_dashboard')).toBeInTheDocument();
@@ -30,7 +27,7 @@ test('renders dashboard chart', async () => {
 test('renders dashboard markdown content', async () => {
   const project = getProject([{ width: 1, markdown: "First Markdown" }, { width: 2, markdown: "Wider Second Markdown" }])
 
-  render(<Dashboard project={project} fetchTraces={fetchTraces} dashboardName={'dashboard'} />)
+  render(<Dashboard project={project} dashboardName={'dashboard'} />)
   const text = await screen.findByText(/First Markdown/);
   expect(text).toBeInTheDocument();
 })
@@ -40,6 +37,6 @@ test('throws when dashboard not found', async () => {
   jest.spyOn(console, 'error').mockImplementation(() => null);
 
   expect(
-    () => render(<Dashboard project={project} fetchTraces={fetchTraces} dashboardName={'noDashboard'} />)
+    () => render(<Dashboard project={project} dashboardName={'noDashboard'} />)
   ).toThrow();
 })
