+++ viewer/src/components/items/Chart.test.js
@@ -1,34 +1,15 @@
-import { render, screen, waitFor } from '@testing-library/react';
+import { screen, waitFor } from '@testing-library/react';
 import Chart from './Chart';
-import { FetchTraceQueryProvider } from '../../contexts/FetchTraceQueryContext';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { renderWithProviders } from '../../utils/test-utils';
 let chart;
 
 beforeEach(() => {
   chart = { name: "name", traces: [] }
 });
 
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      retry: false,
-    },
-  },
-})
 
 test('renders chart', async () => {
-  const fetchTraceQuery = (projectId, name) => ({
-    queryKey: ['trace', projectId, name],
-    queryFn: async () => [],
-  })
-
-  render(
-    <FetchTraceQueryProvider value={fetchTraceQuery}>
-      <QueryClientProvider client={queryClient}>
-        <Chart chart={chart} project={{ id: 1 }} />
-      </QueryClientProvider>
-    </FetchTraceQueryProvider>
-  );
+  renderWithProviders(<Chart chart={chart} project={{ id: 1 }} />);
 
   await waitFor(() => {
     expect(screen.getByText('Mock Plot')).toBeInTheDocument();
