--- viewer/src/contexts/FetchTraceQueryContext.js
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-
-const FetchTraceQueryContext = createContext();
-
-export const FetchTraceQueryProvider = FetchTraceQueryContext.Provider
-export default FetchTraceQueryContext
