+++ viewer/src/contexts/FetchTracesQueryContext.js
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+const FetchTracesQueryContext = createContext();
+
+export const FetchTracesQueryProvider = FetchTracesQueryContext.Provider
+export default FetchTracesQueryContext
