+++ viewer/src/hooks/useTracesData.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect, useContext } from "react";
-import FetchTraceQueryContext from "../contexts/FetchTraceQueryContext";
+import FetchTracesQueryContext from "../contexts/FetchTracesQueryContext";
 import { useQuery } from '@tanstack/react-query'
 import { fetchTracesData } from "../queries/tracesData";
 
 export const useTracesData = (projectId, traceNames) => {
-    const fetchTraceQuery = useContext(FetchTraceQueryContext)
+    const fetchTraceQuery = useContext(FetchTracesQueryContext)
     const [traceData, setTraceData] = useState(null)
 
     const { data: traces } = useQuery(fetchTraceQuery(projectId, traceNames))
