+++ viewer/src/queries/traces.js
@@ -1,6 +1,6 @@
 import { fetchTraces } from '../api/traces'
 
-export const fetchTraceQuery = (projectId, name) => ({
-    queryKey: ['trace', projectId, name],
-    queryFn: async () => fetchTraces(projectId, [name]),
+export const fetchTracesQuery = (projectId, names) => ({
+    queryKey: ['trace', projectId, names],
+    queryFn: async () => fetchTraces(projectId, names),
 })
